Remove duplicate files and fix folders
