+++ src/components/NewBookForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
-import { addBook } from '../redux/booksSlice';
+import { addBook } from '../redux/books/booksSlice';
 
 const categories = [
   'Action',
