--- src/components/redux/store.js
@@ -1,10 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './booksSlice';
-
-const store = configureStore({
-  reducer: {
-    books: booksReducer,
-  },
-});
-
-export default store;
