+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import booksReducer from './booksSlice';
+import booksReducer from './books/booksSlice';
 
 const store = configureStore({
   reducer: {
